Fix markdown for .env section
Due to consecutive markdown blocks not being rendered correctly inside lists (only when accessing the page from the Pycord Guide), I removed one of the two examples that are basically showing the same thing in the same place. In addition to that, I improved the explanation so that it is more cleear for beginners.